Initial commit
This is the initial commit of how I think this mod could look like.
In this version "standard" css files are used. The css file should be
named stylesheet_colors.css, and have a format like this
```
body {
/* Body Text Color */
color: #000000;
/* Body Background Color */
background-color: #ffffff;
}
a {
/* A Link Color */
color: #05a5cb;
}
```
This are stil very crude, and can be refined. Functions are still in the
main file, and not all buttons are working at this time, but I wanted to
give an idea of my vision.